+++ R/getToaYaps.R
@@ -25,14 +25,18 @@ getToaYaps <- function(synced_dat, hydros, rbi_min, rbi_max){
 	ts_focal <- as.matrix(as.data.frame(ts_focal))
 	ts_focal <- ts_focal/10
 	toa <- ts_focal
-	dim(toa)
+
 	dimnames(toa) <- NULL
 
 	# remove rows with too short BI...
 	top1 <- rowMeans(toa, na.rm=TRUE)
 	diffs1 <- c(diff(top1),NA)
 	nobs <- apply(toa, 1, function(k) sum(!is.na(k)))
-	rem_idx <- which(diffs1 < rbi_min-1) # THIS NEEDS TO BE SET BASED ON USED SYSTEM - 1 IS TOO HIGH FOR HR-LIKE
+	if(rbi_min > 10){														### USE PING_TYPE INSTEAD!!!!
+		rem_idx <- which(diffs1 < rbi_min-1) # THIS NEEDS TO BE SET BASED ON USED SYSTEM - 1 IS TOO HIGH FOR HR-LIKE
+	} else {
+		rem_idx <- which(diffs1 < rbi_min-0.1) # rbi_min < 10 should always be HR???
+	}
 	if(length(rem_idx) > 0){
 		toa[rem_idx, ] <- NA
 	}
@@ -47,6 +51,12 @@ getToaYaps <- function(synced_dat, hydros, rbi_min, rbi_max){
 	nobs <- apply(toa, 1, function(k) sum(!is.na(k)))
 	first_ping <- 	which(nobs >= 2)[1]
 	last_ping <- 	rev(which(nobs >= 2))[1]
+	
+	if(is.na(first_ping) | is.na(last_ping) | first_ping == last_ping){
+		print("FATAL ERROR: Not enough data to produce toa")
+		return(FALSE)
+	}
+	
 	toa <- toa[first_ping:last_ping, ]
 	
 	# remake toa-matrix to include pings missed by all hydros...
@@ -57,8 +67,16 @@ getToaYaps <- function(synced_dat, hydros, rbi_min, rbi_max){
 	pings <- data.table::data.table(top=top2, diff=diffs2)
 	pings[, toa_idx:=1:.N]
 	pings[, ping2next := 1]
-	pings[, next_ping_too_late := diff > rbi_max+1]
-	pings[next_ping_too_late==TRUE, ping2next:=ping2next+round(diff/rbi_max)]
+	if(rbi_max > 15){														### USE PING_TYPE INSTEAD!!!!
+		pings[, next_ping_too_late := diff > rbi_max+1] # same as for rbi_min above when trying to exclude impossible pings
+	} else {
+		pings[, next_ping_too_late := diff > rbi_max+.1]
+	}
+	if(rbi_max > 15){ 														 ### USE PING_TYPE INSTEAD!!!!
+		pings[next_ping_too_late==TRUE, ping2next:=ping2next+round(diff/rbi_max)] 
+	} else {
+		pings[next_ping_too_late==TRUE, ping2next:=round(diff/rbi_max)] # the line above puts in an extra pang for pingType = "sbi"
+	}
 	pings[, ping_idx:=cumsum(c(1,ping2next[-.N]))]
 	toa_all <- matrix(ncol=ncol(toa), nrow=max(pings$ping_idx))
 	toa_all[pings$ping_idx, ] <- toa